requirements_test.txt
coverage==4.3.1
coveralls==1.1
django-html-validator==0.4.7
